api-brain-mapper/requirements.txt
alembic==1.13.2
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
Authlib==1.3.0
bcrypt==4.2.1
blinker==1.8.2
certifi==2024.12.14
cffi==1.16.0
charset-normalizer==3.4.1
click==8.1.7
cryptography==42.0.7
distlib==0.3.8
filelock==3.15.4
Flask==3.0.0
Flask-Bcrypt==1.0.1
Flask-Cors==4.0.0
flask-marshmallow==1.2.1
Flask-Migrate==4.0.7
Flask-SQLAlchemy==3.1.1
gevent==24.11.1
greenlet==3.1.1
gunicorn==23.0.0
idna==3.10
itsdangerous==2.1.2
Jinja2==3.1.3
Mako==1.3.5
MarkupSafe==2.1.3
marshmallow==3.21.3
marshmallow-sqlalchemy==1.0.0
minio==7.2.13
packaging==24.1
platformdirs==4.1.0
protobuf==4.21.12
psycogreen==1.0.2
psycopg2==2.9.9
pycparser==2.22
pycryptodome==3.21.0
PyJWT==2.8.0
python-dotenv==1.0.0
pytz==2023.3.post1
requests==2.32.3
SQLAlchemy==2.0.30
timedelta==2020.12.3
typing_extensions==4.12.0
urllib3==2.2.3
virtualenv==20.25.0
Werkzeug==3.0.1
zope.event==5.0
zope.interface==7.2